+++ sdmx-dl-provider-ri/src/main/java/internal/util/credentials/WinPasswordVault.java
@@ -3,6 +3,8 @@ package internal.util.credentials;
 import com.github.tuupertunut.powershelllibjava.PowerShell;
 import com.github.tuupertunut.powershelllibjava.PowerShellExecutionException;
 import lombok.NonNull;
+import nbbrd.design.NotThreadSafe;
+import nbbrd.design.StaticFactoryMethod;
 import nbbrd.io.picocsv.Picocsv;
 import nbbrd.io.text.TextParser;
 import nbbrd.io.text.TextResource;
@@ -15,6 +17,7 @@ import java.io.IOException;
 
 import static java.nio.charset.StandardCharsets.UTF_8;
 
+@NotThreadSafe
 public final class WinPasswordVault implements Closeable {
 
     // https://docs.microsoft.com/en-us/uwp/api/windows.security.credentials.passwordcredential
@@ -31,6 +34,7 @@ public final class WinPasswordVault implements Closeable {
         char[] password;
     }
 
+    @StaticFactoryMethod
     public static @NonNull WinPasswordVault open() throws IOException {
         WinPasswordVault result = new WinPasswordVault();
         result.exec(loadCode());
@@ -95,29 +99,33 @@ public final class WinPasswordVault implements Closeable {
         }
     }
 
+    // The underlying implementation fails on add if size exceeds some limit (somewhere between 4096*3 and 4096*4)
+    // Error: Exception calling "Add" with "1" argument(s): "The file size exceeds the limit allowed and cannot be saved. Cannot add credential to Vault"
+    public static final int MAX_PASSWORD_SIZE = 4096 * 3;
+
     private static final TextParser<PasswordCredential> CREDENTIAL_PARSER = Picocsv.Parser
-            .builder(WinPasswordVault::parseCsv)
-            .options(Csv.ReaderOptions.DEFAULT.toBuilder().maxCharsPerField(4096 * 2).build())
+            .builder(WinPasswordVault::parseCredential)
+            .options(Csv.ReaderOptions.DEFAULT.toBuilder().maxCharsPerField(MAX_PASSWORD_SIZE).build())
             .build();
 
-    private static PasswordCredential parseCsv(Csv.Reader reader) throws IOException {
-        if (reader.readLine() && reader.readLine()) {
-            if (reader.readField()) {
-                String resource = reader.toString();
-                if (reader.readField()) {
-                    String userName = reader.toString();
-                    if (reader.readField()) {
-                        String password = reader.toString();
-                        return new PasswordCredential(resource, userName, password.toCharArray());
-                    }
-                }
-            }
-        }
-        throw new IOException("Invalid content");
+    private static PasswordCredential parseCredential(Csv.Reader csv) throws IOException {
+        if (!csv.readLine()) throw new IOException("Missing CSV header line");
+        if (!csv.readLine()) throw new IOException("Missing CSV data line");
+
+        if (!csv.readField()) throw new IOException("Missing 'Resource' field in credential data");
+        String resource = csv.toString();
+
+        if (!csv.readField()) throw new IOException("Missing 'UserName' field in credential data");
+        String userName = csv.toString();
+
+        if (!csv.readField()) throw new IOException("Missing 'Password' field in credential data");
+        String password = csv.toString();
+
+        return new PasswordCredential(resource, userName, password.toCharArray());
     }
 
     private static String[] loadCode() throws IOException {
-        try (BufferedReader reader = TextResource.newBufferedReader(WinPasswordVault.class, "WinPasswordVault.ps1", UTF_8.newDecoder())) {
+        try (BufferedReader reader = TextResource.newBufferedReader(WinPasswordVault.class, "WinPasswordVault.ps1", UTF_8)) {
             return reader.lines().toArray(String[]::new);
         }
     }